Moving Logging to esp-js so it's consistent across all packages
