+++ packages/esp-js-ui/src/core/index.ts
@@ -3,7 +3,6 @@ export * from './observableExt';
 export * from './decimal';
 export * from './decimalFormat';
 export * from './environment';
-export * from './logger';
 export * from './schedulerService';
 export * from './serialDisposable';
 export * from './unit';
