+++ packages/esp-js/src/system/index.ts
@@ -1,4 +1,4 @@
-import * as logging from './logging';
+import * as logging from './logger';
 import * as utils from './utils';
 import * as disposables from './disposables';
