--- packages/esp-js/src/system/logging/logEvent.ts
@@ -1,8 +1,0 @@
-import {Level} from './level';
-
-export interface LogEvent {
-    logger: string;
-    level: Level;
-    message: string;
-    args: any[];
-}
