--- packages/esp-js/src/system/logging/sink.ts
@@ -1,3 +1,0 @@
-import {LogEvent} from './logEvent';
-
-export type Sink = (logEvent: LogEvent) => void;
