+++ packages/esp-js/src/system/utils.ts
@@ -50,3 +50,7 @@ export function getPropertyNames(object: any): string[] {
     }
     return props;
 }
+
+export const isObject = (value: any): value is object => {
+    return typeof value === 'object' && value !== null;
+};
